+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 - `redeemCheck` params will check if passed `feeCoinSymbol` is equal to `gasCoin` from check, may be omitted, `gasCoin` from check will be used
 - add `maxSupply` field to `CreateCoinTxParams`
 
+## 0.25.0 - 2019-12-25
+- **BREAKING** Change: `decodeLink` now decode `txData` too
+- Add: `getTxData` method which return TxData constructor by txType
+- Fix: `decodeLink` now decodes `nonce` no number instead of string
+
 ## 0.24.1 - 2019-12-24
 - Fix: unbond tx-data